+++ test/WalletFramework.Integration.Tests/WalletFramework.Integration.Tests/WalletFramework.Integration.Tests.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsPackable>false</IsPackable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="FluentAssertions" Version="6.12.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+        <PackageReference Include="Moq" Version="$(MoqVersion)" />
+        <PackageReference Include="xunit" Version="2.9.0" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="coverlet.collector" Version="6.0.2">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\WalletFramework.MdocLib.Tests\WalletFramework.MdocLib.Tests.csproj" />
+      <ProjectReference Include="..\..\WalletFramework.MdocVc.Tests\WalletFramework.MdocVc.Tests.csproj" />
+      <ProjectReference Include="..\..\WalletFramework.Oid4Vc.Tests\WalletFramework.Oid4Vc.Tests.csproj" />
+      <ProjectReference Include="..\..\WalletFramework.SdJwtVc.Tests\WalletFramework.SdJwtVc.Tests.csproj" />
+    </ItemGroup>
+</Project>
