+++ test/WalletFramework.MdocLib.Tests/WalletFramework.MdocLib.Tests.csproj
@@ -13,8 +13,11 @@
     <ItemGroup>
         <PackageReference Include="FluentAssertions" Version="6.12.0"/>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
-        <PackageReference Include="coverlet.collector" Version="6.0.0"/>
-        <PackageReference Include="xunit" Version="2.8.0"/>
+        <PackageReference Include="coverlet.collector" Version="6.0.2">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="xunit" Version="2.9.0" />
     </ItemGroup>
 
     <ItemGroup>
