+++ test/WalletFramework.MdocVc.Tests/WalletFramework.MdocVc.Tests.csproj
@@ -14,7 +14,7 @@
         <PackageReference Include="NUnit" Version="3.13.3"/>
         <PackageReference Include="NUnit3TestAdapter" Version="4.2.1"/>
         <PackageReference Include="NUnit.Analyzers" Version="3.3.0"/>
-        <PackageReference Include="xunit" Version="2.8.1"/>
+        <PackageReference Include="xunit" Version="2.9.0" />
     </ItemGroup>
 
     <ItemGroup>
